+++ packages/frontend/src/Layout/Nav/Nav.types.ts
@@ -2,15 +2,7 @@ export interface HamburgerInnerProps {
   isOpen: boolean;
 }
 
-export interface DrawerMenuProps {
-  isOpen: boolean;
-}
-
 export interface LinkProps {
   $isActive?: boolean;
   $isDesktop?: boolean;
-}
-
-export interface LogoProps {
-  $isDrawerMenuLogo?: boolean;
 }
