+++ packages/frontend/src/components/DrawerMenu/DrawerMenu.types.ts
@@ -1,0 +1,8 @@
+export interface DrawerMenuProps {
+  isOpen: boolean;
+  onClick: () => void;
+}
+
+export interface NavStyledProps {
+  isOpen: boolean;
+}
