+++ packages/frontend/src/enums/fontsize.enum.ts
@@ -2,6 +2,7 @@ export enum FontSize {
   SmallS = "12px",
   Small = "16px",
   Medium = "22px",
+  MediumL = "24px",
   Big = "28px",
   BigL /* https://www.youtube.com/watch?v=qmj1q67NDAk RIP[*] */ = "52px",
 }
