+++ packages/frontend/src/enums/fontweight.enum.ts
@@ -1,0 +1,7 @@
+export enum FontWeight {
+  Light = "300",
+  Regular = "400",
+  Medium = "500",
+  SemiBold = "600",
+  Bold = "700",
+}
